Create seperate file for stateForm
